requirements.txt
flask
flask_sqlalchemy
flask_login
flask_wtf
python-dotenv
email-validator
gunicorn